chore: define option struct fields
Introduce new fields on the Options struct to store the string rep
of the flag content the user should pass to the exec of lilly to invoke
the printing of the version it was built with.

Currently unsure how to handle passing multiple display flags, such as
if the user passes --help --version, not sure what to do, I suppose the
most sensible behaviour would be to just print out both version + help?