Feat/implement show version flag (#181)
* chore: define option struct fields

Introduce new fields on the Options struct to store the string rep
of the flag content the user should pass to the exec of lilly to invoke
the printing of the version it was built with.

Currently unsure how to handle passing multiple display flags, such as
if the user passes --help --version, not sure what to do, I suppose the
most sensible behaviour would be to just print out both version + help?

* refactor: use string builder over long ass string

We would like a helpful function that takes an option and prints
all of the currently supported flags and their behaviour to actually
be useful. The string it returns was becoming unweidly and unbecoming
of a useful function so we've now switched to using a string builder
instead as this lets us nicely lay out the building of the output string

* refactor: adjust location of gitcommit hash resolve

In order to allow the main function body to access the current Lilly
version, the compile time constant resolve has been moved from the
splashscreen module to the root/main and is now just passed in.

* docs: add note comment

The note indicates intentions regarding behaviour handling when the user
potentially passes both the --help and --version flags at the same time
and what to do in this situation.