Feat/render changed cells (#371)
* fix: change cell iteration by grid width and height

* feat: naive attempt to do only changed cell re-render pass

* temp: use new ctx and document view

* swap back to old document renderer

* fix: checking to render on just change is no longer necessary

This flag was very useful for at least reducing the amount of flickering
or stopping it from happening if the UI had definitely not been changed,
but now it isn't needed, given that we have the same behaviour from
immediate ctx.

* experiment: for now calling clear is still required, unsure why

Not sure why I can't just make the clear be the last step of the last
flush, I think there's a good reason... ah yeah if we don't use it as
its own call then the very first frame would result in nothing being
rendered ever.

* fix: just stop trying to virtualise the cursor, give up

Stop it. Get some help.

* fix: actually pay attention to hide cursor directive

* expriment: invoke hide cursor to reduce headaches for tauraamui

* chore: flip back to using old ctx for now to push to master safely

* test: ensure default grid size is 100

* fix: stop resizing down panicking

This fixes lilly crashing when resizing from a large grid to a smaller
one. It also prevents the initial frame being 1x1 making the grid
initialiser unhappy.