feat: pass in syntax defs to rendering parser
These changes share the language syntax defs with the buffer view
renderers syntax parser, the aim being to use this to adjust the
returned type of each token based on its presence within the "keywords"
set of the languages def or not.

We're also changing the behaviour of the syntax parser to be a
consistent instance but reset each iter so we don't have to re-assign
the same static field each time.