LIL-414 Refactor visual line selection highlight for buffer view (#458)
* refactor: move fg color resolution into own function

* fix: remove unwanted bg reset

* test: start converting existing buffview tests to clearer

This will be easier to ensure that each token rendered is using the
correct fg and bg colors in the place of use.

* test: convert existing test to use new colour drawn text checking

* fix: not ideal but guard reset background reset with visual line check

* test: replace draw line 1 test mockctx with mockcolorctx

* test: move remaining bufferview tests to use coloredmockctx

* test: note to indicate we now have enough tests prior to further work

* refactor: simplify current full visual line selection highlight

Make the tokens in charge of checking if they're within a selection, in
case of visual line we just check if the selection is "full" for that
given line.