refactor: rename gap_buf data cursor methods to be clearer
There is a concept of a "data cursor" which is essentially just where
the current gap buffer gap start/insertion point is, this is a separate
concept to the cursor which denotes where in the "document data"
abstraction new data should be inserted.

Until this rename I was getting sometimes lost as to which cursor type
these methods related to, data layer or document layer cursors.